feat: setup layered defines config
