+++ config/defines/base-env.ts
@@ -1,0 +1,28 @@
+import { getEnvironment } from '../../envs/getEnvs';
+import { SDK_VERSION } from '../version';
+
+export const getBaseEnvDefines = (mode: string) => {
+    const { api, app } = getEnvironment(mode);
+    const { session } = api;
+
+    const isDevelopment = mode === 'development';
+    const isTestEnv = process.env.TEST_ENV === '1';
+    const localAssets = app.useCdn === 'true' ? null : isDevelopment || isTestEnv;
+    const testCdnAssets = isDevelopment || app.useTestCdn === 'true' ? true : null;
+
+    return {
+        'process.env.SDK_VERSION': JSON.stringify(SDK_VERSION),
+        'process.env.SESSION_ACCOUNT_HOLDER': JSON.stringify(session.accountHolder || null),
+        'process.env.SESSION_AUTO_REFRESH': JSON.stringify(isDevelopment ? session.autoRefresh === 'true' || null : undefined),
+        'process.env.SESSION_MAX_AGE_MS': JSON.stringify(isDevelopment ? session.maxAgeMs || null : undefined),
+        'process.env.SESSION_PERMISSIONS': JSON.stringify(session.permissions || null),
+        'process.env.TEST_ENV': JSON.stringify(process.env.TEST_ENV),
+        'process.env.USE_CDN': JSON.stringify(app.useCdn ?? null),
+        'process.env.VITE_APP_LOADING_CONTEXT': JSON.stringify(isDevelopment ? app.loadingContext || null : null),
+        'process.env.VITE_APP_PORT': JSON.stringify(app.port || null),
+        'process.env.VITE_APP_URL': JSON.stringify(process.env.DEPLOY_PRIME_URL?.replace('main--', '') || app.url || null),
+        'process.env.VITE_LOCAL_ASSETS': JSON.stringify(localAssets),
+        'process.env.VITE_MODE': JSON.stringify(process.env.VITE_MODE ?? mode),
+        'process.env.VITE_TEST_CDN_ASSETS': JSON.stringify(testCdnAssets),
+    };
+};
